News Working
There is now a news system. Whenever the user starts a new day, a random headline will be selected. It will be checked to make sure it is not already active. It will then proceed to change the news related impact stats of the stock. The stock price will only be affected the next day. The day the headline comes out is the the day before the affect will be seen. This will give time for the user to make changes to their positions before the impact.