News Further Implemented
I have further added the news feature. Now, the news has an effect on the stocks, either up or down. Also, if there is already news tied to a certain stock, and a new piece is selected that affects that same stock, it will not be chosen. There were also several changes including mispelled names, and other various fixes.